# gen/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email:   
    name:     
    url: https://josue-lubaki.ca
  description: Springular REST API
  license:
    name: Apache License Version 2.0
  title: Springular REST API
  version: 1.0.0
servers:
- description: Base Url
  url: http://localhost:8080
...
paths:
...
  /api/v1/haircuts/{id}:
    get:
      operationId: getHaircutUsingGET
      parameters:
      - description: Identifier of the hairstyle to consult
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HaircutDTO'
          description: Successfully retrieved haircut
      summary: Get haircut by id
      tags:
      - Haircut consultation
      x-accepts: application/json
  /management/api/v1/haircuts:
    get:
      operationId: getHaircutsUsingGET_1
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/HaircutDTO'
                type: array
          description: OK
        "400":
          description: Bad Request
      security:
      - JWT:
        - global
      summary: Get all haircuts
      tags:
      - Haircut Management Controller
      x-accepts: '*/*'
    post:
      operationId: createHaircutUsingPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HaircutDTO'
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/HaircutDTO'
          description: OK
        "201":
          description: Created
        "400":
          description: Bad Request
      security:
      - JWT:
        - global
      summary: Create haircut
      tags:
      - Haircut Management Controller
      x-contentType: application/json
      x-accepts: '*/*'
  /management/api/v1/haircuts/{id}:
    delete:
      operationId: deleteHaircutUsingDELETE
      parameters:
      - description: id
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Deleted hair
        "400":
          description: Bad Request
      security:
      - JWT:
        - global
      summary: Delete haircut
      tags:
      - Haircut Management Controller
      x-accepts: application/json
    get:
      operationId: getHaircutUsingGET_1
      parameters:
      - description: id
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HaircutDTO'
          description: OK
        "400":
          description: Bad Request
      security:
      - JWT:
        - global
      summary: Get haircut by ID
      tags:
      - Haircut Management Controller
      x-accepts: application/json
    put:
      operationId: updateHaircutUsingPUT
      parameters:
      - description: id
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HaircutDTO'
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/HaircutDTO'
          description: Updated hair
        "400":
          description: Bad Request
      security:
      - JWT:
        - global
      summary: Update haircut
      tags:
      - Haircut Management Controller
      x-contentType: application/json
      x-accepts: '*/*'
...
components:
  schemas:
...
    HaircutDTO:
      example:
        estimatedTime: estimatedTime
        price: 0
        imageURL: imageURL
        description: description
        id: id
        title: title
      properties:
        description:
          type: string
        estimatedTime:
          type: string
        id:
          type: string
        imageURL:
          type: string
        price:
          format: int32
          type: integer
        title:
          type: string
      title: HaircutDTO
      type: object
...
  securitySchemes:
    JWT:
      in: header
      name: JWT
      type: apiKey